requirements.txt
mitmproxy==2.0.2
pytest==3.3.1
pytest-tornado==0.4.5
